% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BANEScarparkinglite.R
\docType{package}
\name{BANEScarparkinglite-package}
\alias{BANEScarparkinglite-package}
\title{Functions for obtaining and working with car parking data from Bath and North East Somerset}
\description{
Contains functions for importing and working with the BANES car parking
 records and other related datasets. For the full version of the package, including
 all datasets, see the repo at https://github.com/owenjonesuob/BANEScarparking.
}
\author{
Owen Jones (\email{olj23@bath.ac.uk})
}
