% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_records.R
\name{get_all_crude}
\alias{get_all_crude}
\title{Download all raw records from Bath: Hacked datastore}
\usage{
get_all_crude()
}
\value{
The full dataset of car parking records
}
\description{
Reads the entire CSV file found on the Bath: Hacked datastore
 (\url{http://bit.ly/2i3Y1uF}). The data frame created can subsequently be
 updated using \code{\link{refuel_crude}}.\cr
 \emph{\strong{Warning:} The file is very large! This may take a while to
 run, depending on your internet connection.}
}
\examples{
raw_data <- get_all_crude()

str(raw_data)
}
\seealso{
\itemize{
 \item \code{\link{refuel_crude}} for updating raw records
 \item \code{\link{refuel}} for updating data already processed with
  \code{\link{refine}}
}
}
