% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_records.R
\name{refuel_crude}
\alias{refuel_crude}
\title{Add new raw records from Bath: Hacked datastore}
\usage{
refuel_crude(x)
}
\arguments{
\item{x}{The data frame of raw records from the Bath: Hacked datastore (see
\code{\link{get_all_crude}}).}
}
\value{
The data frame updated with any more recent records.
}
\description{
Update  data frame of raw records with any records that have since been added
 to the Bath: Hacked datastore.
}
\examples{
\donttest{
raw_data <- get_all_crude()

# Some time later...
raw_data <- refuel_crude(raw_data)
}
}
\seealso{
\itemize{
 \item \code{\link{get_all_crude}} for obtaining data frame of raw records
 \item \code{\link{refuel}} for updating data already processed with
  \code{\link{refine}}
}
}
